functions/ingestion/requirements.txt
functions-framework==3.3.0
numpy==1.24.3
pandas==2.0.3
google-cloud-storage==2.7.0